refactor: move custom commands validation to shared lib
Moves validation logic and types to @shared/lib/custom-commands:
- CustomProjectCommand interface
- ValidationResult interface
- validateCustomCommand function
- replaceTemplateVariables function

Fixes cross-layer import violation (main importing from renderer).
Renderer lib now re-exports from shared for backward compatibility.
